requirements.txt
JayDeBeApi==1.2.3
requests==2.31.0
pyyaml==6.0
